refactor(shopify): split index.ts monolith into focused modules
Break the 440-line lib/integrations/shopify/index.ts into client.ts,
products.ts, collections.ts, cart-operations.ts, pages.ts; index.ts is now a
thin barrel (export *) so consumer imports are unchanged. Extract the shared
reshape helpers (removeEdgesAndNodes, reshapeImages, reshapeProduct,
reshapeProducts) into reshape.ts instead of duplicating them per module.
Behavior-preserving (#95 / roadmap Phase 2).
